requirements.txt
binance-connector==3.6.0
python-dotenv==1.0.0
requests==2.31.0
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
numpy==1.24.3
python-socketio==5.9.0
python-engineio==4.7.1
APScheduler==3.10.4

